[TEST] rename test/unit/io/{alignment => sam}_file
